This program assumes that the arrivals and service occur according to exponential distribution.
1. Firstly, I have assumed that the inter arrival times and the service
times are exponentially distributed.
2. Then the program asks for the mean values of the service and the
inter arrival times.
3. Then the user gets to choose the method of simulation:
a. Based on number of runs  or
b. Based on total time duration
4. The trials case:
a. The user enters the no. of trials and no. of runs in each trial.
b.  The total runs can range from few hundreds (500s) to 5000s
c. The performance measures are displayed.
d. The user can also compare these values with the steady state
parameters.
5. The total duration to run case:
a. The user enters the total time to run.
b. The performance measures are displayed.